fuck

+++ Win32Project1/Win32Project1/BaseObject2D.h
@@ -26,7 +26,7 @@ public:
 	int Fadeout = -1; // フェードアウトのフレームを算出
 	int vTransparency = -1; // 透過度変化
 
-	int GrazeScore; // かすった時に加算されるスコア
-	int EraseScore; // ボムで消える時に加算されるスコア
+	double GrazeScore; // かすった時に加算されるスコア
+	double EraseScore; // ボムで消える時に加算されるスコア
 };
